[NFC] Cleanup for compiler warnings
Many NFC (hopefully) changes, to resolve warnings for:

* Unused variables
* Deprecated functions
    * type->getPointerTo(ASpace) =>
      llvm::PointerType::get(type->getContext(), ASpace)
    * Intrinsic::getDeclaration =>
      Intrinsic::getOrInsertDeclaration
* Unknown escape sequence "\/"
* `default:` switch target for exhaustive switch

I don't believe any of these are applicable upstream, or in the case of
at least one of the unittest fixes the change came direclty from
upstream and simply hadn't made it to us somehow.

+++ clang/unittests/libclang/LibclangTest.cpp
@@ -1220,7 +1220,7 @@ static_assert(true, message);
   const char *Args[] = {"-xc++", "-std=c++26"};
   ClangTU = clang_parseTranslationUnit(Index, fileName.c_str(), Args,
                                        std::size(Args), nullptr, 0, TUFlags);
-  ASSERT_EQ(clang_getNumDiagnostics(ClangTU), 0);
+  ASSERT_EQ(clang_getNumDiagnostics(ClangTU), 0u);
   std::optional<CXCursor> staticAssertCsr;
   Traverse([&](CXCursor cursor, CXCursor parent) -> CXChildVisitResult {
     if (cursor.kind == CXCursor_StaticAssert) {
@@ -1229,7 +1229,7 @@ static_assert(true, message);
     return CXChildVisit_Continue;
   });
   ASSERT_TRUE(staticAssertCsr.has_value());
-  size_t argCnt = 0;
+  int argCnt = 0;
   Traverse(*staticAssertCsr, [&argCnt](CXCursor cursor, CXCursor parent) {
     switch (argCnt) {
     case 0:
